part2_textSpilt_graphGenerate/requirements.txt
langchain_community>=0.0.13
langchain_core>=0.1.9
google-generativeai>=0.3.1
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0
zhipuai>=2.0.1
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0 